+++ lib/asn1.dart
@@ -5,18 +5,18 @@ export 'asn1/asn1_object.dart';
 export 'asn1/asn1_parser.dart';
 export 'asn1/asn1_tags.dart';
 export 'asn1/asn1_utils.dart';
-export 'asn1/objects/asn1_bit_string.dart';
-export 'asn1/objects/asn1_boolean.dart';
-export 'asn1/objects/asn1_enumerated.dart';
-export 'asn1/objects/asn1_ia5_string.dart';
-export 'asn1/objects/asn1_integer.dart';
-export 'asn1/objects/asn1_null.dart';
-export 'asn1/objects/asn1_object_identifier.dart';
-export 'asn1/objects/asn1_octet_string.dart';
-export 'asn1/objects/asn1_printable_string.dart';
-export 'asn1/objects/asn1_sequence.dart';
-export 'asn1/objects/asn1_set.dart';
-export 'asn1/objects/asn1_utc_time.dart';
-export 'asn1/objects/asn1_utf8_string.dart';
+export 'asn1/primitives/asn1_bit_string.dart';
+export 'asn1/primitives/asn1_boolean.dart';
+export 'asn1/primitives/asn1_enumerated.dart';
+export 'asn1/primitives/asn1_ia5_string.dart';
+export 'asn1/primitives/asn1_integer.dart';
+export 'asn1/primitives/asn1_null.dart';
+export 'asn1/primitives/asn1_object_identifier.dart';
+export 'asn1/primitives/asn1_octet_string.dart';
+export 'asn1/primitives/asn1_printable_string.dart';
+export 'asn1/primitives/asn1_sequence.dart';
+export 'asn1/primitives/asn1_set.dart';
+export 'asn1/primitives/asn1_utc_time.dart';
+export 'asn1/primitives/asn1_utf8_string.dart';
 export 'asn1/unsupported_asn1_encoding_rule_exception.dart';
 export 'asn1/unsupported_asn1_tag_exception.dart';
