+++ lib/asn1/primitives/asn1_enumerated.dart
@@ -1,7 +1,7 @@
 import 'dart:typed_data';
 
 import 'package:pointycastle/asn1/asn1_tags.dart';
-import 'package:pointycastle/asn1/objects/asn1_integer.dart';
+import 'package:pointycastle/asn1/primitives/asn1_integer.dart';
 
 ///
 /// An ASN1Enumerated object
