+++ test/asn1/primitives/asn1_ia5_string_test.dart
@@ -2,7 +2,7 @@ import 'dart:typed_data';
 
 import 'package:pointycastle/asn1/asn1_encoding_rule.dart';
 import 'package:pointycastle/asn1/asn1_tags.dart';
-import 'package:pointycastle/asn1/objects/asn1_ia5_string.dart';
+import 'package:pointycastle/asn1/primitives/asn1_ia5_string.dart';
 import 'package:test/test.dart';
 
 void main() {
