+++ test/asn1/primitives/asn1_null_test.dart
@@ -1,7 +1,7 @@
 import 'dart:typed_data';
 
 import 'package:pointycastle/asn1/asn1_encoding_rule.dart';
-import 'package:pointycastle/asn1/objects/asn1_null.dart';
+import 'package:pointycastle/asn1/primitives/asn1_null.dart';
 import 'package:test/test.dart';
 
 void main() {
